working simple strat

+++ src/types/Point2D.cpp
@@ -1,7 +1,7 @@
 #include <types/Point2D.h>
 #define UNIT 1
 
-Point2D::Point2D(int x, int y)
+Point2D::Point2D(float x, float y)
 {
 	X = x;
 	Y = y;
